# survey-parent/api-survey/src/main/resources/static/json/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "survey",
    "version": "v1.0",
    "title": "API文档",
    "termsOfService": "lego.cn",
    "contact": {
      "name": "lego",
      "url": "lego.cn",
      "email": "survey@legocloud.cn"
    }
  },
  "host": "192.168.101.197:50010",
  "basePath": "/",
...
  "paths": {
...
    "/survey/api-report/excelReport/generate": {
      "post": {
        "tags": [
          "excel-report-controller"
        ],
        "summary": "生成excel测量成果报表",
        "description": "生成excel测量成果报表",
        "operationId": "generateReportUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "name": "endTimestamp",
            "in": "query",
            "description": "endTimestamp",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "pageIndex",
            "in": "path",
            "description": "当前页",
            "required": false,
            "type": "integer",
            "default": 1,
            "format": "int32",
            "x-example": 1
          },
          {
            "name": "pageSize",
            "in": "path",
            "description": "每页大小",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32",
            "x-example": 10
          },
          {
            "name": "sectionId",
            "in": "path",
            "description": "标段id",
            "required": true,
            "type": "string"
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "name": "startTimestamp",
            "in": "query",
            "description": "startTimestamp",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          },
          {
            "name": "workspaceId",
            "in": "query",
            "description": "workspaceId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
    "/survey/api-report/excelReport/upload/{sectionId}": {
      "post": {
        "tags": [
          "excel-report-controller"
        ],
        "summary": "上传excel数据",
        "description": "上传excel数据",
        "operationId": "uploadReportUsingPOST",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "name": "reportFile",
            "in": "formData",
            "description": "reportFile",
            "required": false,
            "type": "file"
          },
          {
            "name": "sectionId",
            "in": "path",
            "description": "标段id",
            "required": true,
            "type": "string"
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
...
    "/survey/api-user/config/create": {
      "post": {
        "tags": [
          "config-controller"
        ],
        "summary": "添加配置信息",
        "description": "添加配置信息",
        "operationId": "addUsingPOST_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "config",
            "description": "config",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Config"
            }
          },
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO"
            }
          },
          "404": {
            "description": "Resources Not Found",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
    "/survey/api-user/config/delete": {
      "delete": {
        "tags": [
          "config-controller"
        ],
        "summary": "删除配置信息",
        "description": "删除配置信息",
        "operationId": "deleteUsingDELETE_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "name": "id",
            "in": "query",
            "description": "配置id",
            "required": true,
            "type": "string"
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO"
            }
          },
          "404": {
            "description": "Resources Not Found",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
    "/survey/api-user/config/info": {
      "get": {
        "tags": [
          "config-controller"
        ],
        "summary": "查询配置信息",
        "description": "查询配置信息",
        "operationId": "infoUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "name": "id",
            "in": "query",
            "description": "配置id",
            "required": true,
            "type": "string"
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO"
            }
          },
          "404": {
            "description": "Resources Not Found",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
    "/survey/api-user/config/list/{pageIndex}": {
      "get": {
        "tags": [
          "config-controller"
        ],
        "summary": "查询配置信息",
        "description": "查询配置信息",
        "operationId": "listUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "name": "pageIndex",
            "in": "path",
            "description": "当前页",
            "required": true,
            "type": "integer",
            "format": "int32",
            "x-example": 1
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页大小",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32",
            "x-example": 10
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO"
            }
          },
          "404": {
            "description": "Resources Not Found",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
    "/survey/api-user/config/modify": {
      "put": {
        "tags": [
          "config-controller"
        ],
        "summary": "修改配置信息",
        "description": "修改配置信息",
        "operationId": "modifyUsingPUT_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "config",
            "description": "config",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Config"
            }
          },
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO"
            }
          },
          "404": {
            "description": "Resources Not Found",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
    "ApiError": {
      "type": "object",
      "title": "ApiError"
    },
...
    "Config": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "配置id"
        },
        "name": {
          "type": "string",
          "description": "配置名称"
        },
        "option": {
          "type": "array",
          "description": "可选项",
          "items": {
            "type": "string"
          }
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间"
        },
        "valid": {
          "type": "integer",
          "format": "int32",
          "description": "是否有效 0-是;1-否"
        }
      },
      "title": "Config"
    },
...
    "RespVO": {
      "type": "object",
      "properties": {
        "info": {
          "type": "object"
        },
        "msg": {
          "type": "string"
        },
        "ret": {
          "type": "string"
        },
        "retCode": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "RespVO"
    },
...
  }
}